+++ lib/output/__init__.py
@@ -1,2 +1,2 @@
-from .cli_output import CLIOutput  # noqa: F401
-from .print_output import PrintOutput  # noqa: F401
+from .verbose_output import CLIOutput  # noqa: F401
+from .silent_output import PrintOutput  # noqa: F401
