+++ lib/output/silent_output.py
