+++ lib/output/verbose_output.py
